squash xrope.h into string.h and string.c
