+++ extlib/benz/include/picrin.h
@@ -51,7 +51,6 @@ extern "C" {
 #include "picrin/xvect.h"
 #include "picrin/xhash.h"
 #include "picrin/xfile.h"
-#include "picrin/xrope.h"
 
 #include "picrin/value.h"
